Move orswot tests into the /tests directory
